# MRFerreira.API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MRFerreira - Documentação",
        "description": "Documentação de todas as rotas da MRFerreira",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/api/providers": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Listar fornecedores",
                "operationId": "386f3ddc2b7b26effa6a559287ccbe4c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "cnpj": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "phone": {
                                                        "type": "string"
                                                    },
                                                    "cellphone": {
                                                        "type": "string"
                                                    },
                                                    "logo": {
                                                        "type": "string"
                                                    },
                                                    "logo_url": {
                                                        "type": "string"
                                                    },
                                                    "address": {
                                                        "properties": {
                                                            "zipcode": {
                                                                "type": "string"
                                                            },
                                                            "street": {
                                                                "type": "string"
                                                            },
                                                            "neighborhood": {
                                                                "type": "string"
                                                            },
                                                            "number": {
                                                                "type": "string"
                                                            },
                                                            "state": {
                                                                "type": "string"
                                                            },
                                                            "city": {
                                                                "type": "string"
                                                            },
                                                            "complement": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-07-25T15:30:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Cadastrar um novo fornecedor",
                "operationId": "c6b8bd1bd0d9d78359553ea08bf93208",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "logo",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Forncedor 1"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "01234567890123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "012345678901234"
                                    },
                                    "cellphone": {
                                        "type": "string",
                                        "example": "012345678901234"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "address": {
                                        "properties": {
                                            "zipcode": {
                                                "type": "string",
                                                "example": "0123456"
                                            },
                                            "street": {
                                                "type": "string",
                                                "example": "Rua 1"
                                            },
                                            "number": {
                                                "type": "integer",
                                                "example": "123"
                                            },
                                            "neighborhood": {
                                                "type": "string",
                                                "example": "Bairro 1"
                                            },
                                            "state": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "complement": {
                                                "type": "string",
                                                "example": "Complemento 1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "properties": {
                                                "field": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Please provide a valid JWT token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}